requirements.txt
phidata
streamlit
pandas
plotly
nltk
wordcloud
groq
numpy